requirements.txt
PyYAML==6.0.3
streamlit==1.41.1
python-dotenv==1.1.1
pandas==2.3.3
numpy==2.2.6

langchain-openai==1.0.1
langchain-ollama==1.0.0
langchain-groq==1.0.0

langgraph==1.0.1
langchain-core==1.2.3
langchain-community==0.4.1
langchain-text-splitters==1.0.0

ollama==0.6.0
scikit-learn==1.7.2
ipython==8.30.0